+++ source/image-handler/test/semantic-mapper/resize.spec.ts
@@ -19,7 +19,6 @@ describe("resize", () => {
     expect(edits).toEqual(expectedResult.edits);
   });
 
-
   it("Should pass if the proper edit translations are applied and in the correct order", () => {
     // Arrange
     const path = "/test-image-001.jpg?h=300";
@@ -53,6 +52,21 @@ describe("resize", () => {
   it("Should pass if the proper edit translations are applied and in the correct order", () => {
     // Arrange
     const path = "/test-image-001.jpg?w=0&h=0";
+
+    // Act
+    const customMapper = new SemanticMapper();
+    const edits = customMapper.mapPathToEdits(path);
+
+    // Assert
+    const expectedResult = {
+      edits: { resize: { width: null, height: null, fit: "inside" } },
+    };
+    expect(edits).toEqual(expectedResult.edits);
+  });
+
+  it("Tests error on deployment, no parameters produces these edits", () => {
+    // Arrange
+    const path = "/test-image-001.jpg";
 
     // Act
     const customMapper = new SemanticMapper();
